Fixed NaN loss bug

+++ encoder.py
@@ -49,7 +49,7 @@ class DataEncoder:
             grid_size = input_size / fm_size
             fm_w, fm_h = int(fm_size[0]), int(fm_size[1])
             xy = meshgrid(fm_w,fm_h) + 0.5  # [fm_h*fm_w, 2]
-            xy = (xy*grid_size).view(fm_h,fm_w,1,2).expand(fm_h,fm_w,9,2)
+            xy = (xy.float()*grid_size).view(fm_h,fm_w,1,2).expand(fm_h,fm_w,9,2)
             wh = self.anchor_wh[i].view(1,1,9,2).expand(fm_h,fm_w,9,2)
             box = torch.cat([xy,wh], 3)  # [x,y,w,h]
             boxes.append(box.view(-1,4))
